F - Gjør det mulig å resende søknaden fra feil-ved-innsending
