+++ src/pages/feil-ved-innsending/index.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement } from 'react';
+import React, {ReactElement, useContext} from 'react';
 import { BodyLong, Button, Heading } from '@navikt/ds-react';
 import BannerLayout from '@/components/banner-layout/BannerLayout';
 import CustomGuidePanel from '@/components/custom-guide-panel/CustomGuidePanel';
@@ -8,10 +8,19 @@ import defaultValues from '@/defaultValues';
 import { useRouter } from 'next/router';
 import { useFormContext } from 'react-hook-form';
 import Søknad from '@/types/Søknad';
+import {UtfyllingContext} from "@/pages/utfylling/[[...step]]";
+import {InnsendingContext} from "@/pages/_app";
+import {sendInnSøknad} from "@/utils/innsending";
 
 export default function FeilVedInnsending() {
     const router = useRouter();
-    const { reset } = useFormContext<Søknad>();
+    const { reset, getValues } = useFormContext<Søknad>();
+    const { personalia, valgtTiltak } = useContext(UtfyllingContext);
+    const { setSøknadsinnsendingInProgress, setInnsendingstidspunkt, søknadsinnsendingInProgress } = useContext(InnsendingContext);
+
+    function navigateToHome() {
+        return router.push('/', undefined, { shallow: false });
+    }
 
     React.useEffect(() => {
         router.beforePopState(({ as }) => {
@@ -42,10 +51,18 @@ export default function FeilVedInnsending() {
             <div className={styles.lastInnSidenPåNyttButton}>
                 <Button
                     type="button"
-                    onClick={() => {
-                        const basePath = process.env.NEXT_PUBLIC_BASE_PATH ?? '';
-                        (window.location as any) = basePath;
-                    }}
+                    onClick={ () => sendInnSøknad(router, getValues(), personalia!, valgtTiltak!, setSøknadsinnsendingInProgress!, setInnsendingstidspunkt!) }
+                    disabled={søknadsinnsendingInProgress}
+                    loading={søknadsinnsendingInProgress}
+                >
+                    Send inn på nytt
+                </Button>
+            </div>
+            <div className={styles.lastInnSidenPåNyttButton}>
+                <Button
+                    type="button"
+                    variant="secondary"
+                    onClick={navigateToHome}
                 >
                     Start søknaden på nytt
                 </Button>
