+++ src/pages/kvittering/kvittering.module.css
@@ -20,3 +20,15 @@
     display: block;
     width: fit-content;
 }
+
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding-bottom: 2.5rem;
+}
+
+.body {
+    max-width: 600px;
+}
